feat: Replace raw SQLite usage by sqlx
Fix tests

fmt

doc: Update pueue version

remove dbg

doc: Update README

feat: Remove previous Pueue task during start

+++ crates/jocker-lib/src/lib.rs
@@ -11,4 +11,4 @@ pub mod stop;
 
 pub const JOCKER: &str = "jocker";
 
-pub type Pid = u32;
+pub type Pid = usize;
